use non-static sacy_Config
the moment you begin to add multiple {asset_compile} blocks, the static
sacy_Config singleton felt a bit unsave (though it would probably not
have mattered when you consider that PHP is single-threaded).

Now were are creating a new sacy_Config instance for every block and use
that for further operations.

This also led me to create an abstract sacy_ConfiguredRenderHandler that
handles storage of the references to both the smarty and the config
objects - no need to duplicate code there.