rework build script to generate a .phar file
when I was looking at how to add phamlp to the bundle, I stumbled upon
that fact that said library is HUGE and does all kinds of including on
its own.

While I could have either

  a) forked phamlp and stripped it down (ugh)
  b) tried to make the bundling into one file work somehow (ugh)

I decided that just building a phar file and not touching any included
file at all was the easier (probably the only workable one) solution.

Of course this means that we now require PHP 5.3 (for the bundled
solution), but I really think requiring a 2 years old language is
something that can be seen as acceptable)