hook in lessphp support
we let the render handler export a list of supported types and make that
list dependent of the availability of the lessc class. Also we include
lessc.inc.php if it's there (which is the case if we are running from
inside the phar archive)

The cool thing about this is that you can still use the non-bundled
version and just include your own copy of lessc.