don't blow up if there's nothing to do
if you use {asset_compile} around a block of content that doesn't have
any <link> or <script> tag, we would have blown up in the final
unconditional call to renderFiles() due to an empty $entry.

Another cause for this is if {asset_compile} is wrapped around links and
scripts which don't need compilation and debug_mode is 3

The third cause is if {asset_compile} is wrapped around links and
scripts which all point to invalid files.

this fixes gh-5 (finally)