change terminology
in order to move to handling tags with content, let's use the term
"work unit" for the thing sacy works on. This could be a file or a bunch
of content to work with.

Right now, it's just a file. This patch doesn't change the functionality
at all aside of renaming methods and variable.

And moving stuff out of the huge smarty_block_asset_compile function