remove dependencies to smarty
we only really passed around that smarty instance in order to call
trigger_error() on it which is actually not even available in Smarty 3
any more.

By having a smarty agnostic backend (not there yet), sacy could even be
used in a smarty-less environment at some point.

This also gets rid of the remaining two global functions (aside of the
block function of course)