handle work units without files
while we are still not accepting tags with content, we now have the
framework in plave to handle work units that are based on contents, but
not on files.

In order to do change detection, we are using the md5() of the content
to sacify and then use a very simple file-system backed cache (you can
provide a different class using a define() if you have something else
already - like memcache).

Before coming up with this implementation I tried various reworks of the
whole sacy concept, but what ever I came up with: It was never totally
correct (sacy never was - there are issues when mixing sacyfiable and
non-sacyfiable tags), still messy and producing tons and tons of small
objects in the process.

So in the end, I kept the implementation simple