get rid of FragmentCache interface
duck typing seems to be a better method here as having to implement the
interface in the library-using code would seriously miss with smarty's
autoloading ability (i.e. the consumer would have to manually include
the file to make the interface known).

As this is an advanced feature which requires an additional define(), I
think we can live with actually just duck-typing this one