parse tag attributes that contain dashes
The regex is wrong, it will also match attr names like `a-` or `-b?` or
even `-`, which are of course not valid.

Since the whole concept of sacy is centered around parsing
developer-supplied html and doesn't use a full-blown parser, I feel
the compromise is acceptable here in order to keep the regex simple.