convert to Composer
this whole hassle with the phar file is SO not worth it any more in the
age of composer.

This is a pretty big rewrite of most components, but now, sacy can be
used via composer.

It will also not have any more side effects when it's used, so the
smarty function will not be made available by default any more. Instead,
you can create a Sacy instance and then call registerSmartyPlugin() and
pass it a Smarty instance which will then be able to use the plugin.

You can still use registerGlobalCompatHandler in order to install a
globally available Smarty function.

Also, while the old define()'s still exist for configuring sacy, it's
recommended you now pass an instance of a class implementing
Configuration on instance creation. This is more strongly typed and much
clearer what it does.

This explicit configuration, combined with the per-smarty plugin
registration also allows for an unprecedented level of configurability